Created EdgeVerticle2 that supports DeploymentOptions for config
The existing `EdgeVerticle` creates a vertx object in its constructor. This causes the `AbstractVerticle` to lose its vertx, which is supplied by the launch script. The launch vertx has the context where the deployment options are provided to the verticle, so deployment options are not available with `EdgeVerticle`.

To get around this, `EdgeVerticle2` was created. The constructor no longer creates a vertx and all operations are now done in the verticle `start` method using the launch provided vertx instance. The `start` method is provided by the `AbstractVerticle` and we use this method to perform our potentially long setup operations (secure store loading, etc.) that could block the main thread. Other entry points to the verticle behave synchronously, so are not ideal for this purpose.

Default deployment options are loaded by the verticle. These can be overridden by env variable, which can be further be overridden by deployment options provided by the launcher.