Fix for EDGCOMMON-17
Used maven profiles to create OS dependent variables in order to store the assembly regex. This is to overcome the need to escape the '/' using a '\' on Windows to make the assembly plugin work, however, the backslash causes nothing to be included on Unix. Now we can use both and only the right one will be active on the host os.