EDGCOMMON-20: enable HTTP client compression support
Set the `HttpClientOption` to allow the client to request compression be used if supported by the server. This allows pass through compression headers to not cause the server to unexpectedly send compressed data to the client's handlers, since it will now be taken care of by the client prior to reaching the handler with this option set.