EDGCOMMON-28: Upgrade to Java 11
Upgraded the library to build with Java 11. Rest Assured was also
upgraded since the existing version uses a class that is not available
in Java 11. Newer versions no longer use this class. Versions that
include this fix have relocated the package, so imports needed to be
updated. Actual code within tests was not impacted. No code was changed
to take advantage of Java 11 features.